--- packages/style/src/helpers/isAtRule.ts
@@ -1,3 +1,0 @@
-export default function isAtRule(value: string): boolean {
-  return value[0] === '@';
-}
