--- packages/style/src/helpers/isKeyframesRule.ts
@@ -1,4 +1,0 @@
-export default function isKeyframesRule(value: string): boolean {
-  // eslint-disable-next-line no-magic-numbers
-  return value.slice(0, 10) === '@keyframes';
-}
