--- packages/style/src/helpers/isNestedSelector.ts
@@ -1,13 +1,0 @@
-export default function isNestedSelector(value: string): boolean {
-  const char = value[0];
-
-  return (
-    char === ':' ||
-    char === '[' ||
-    char === '>' ||
-    char === '~' ||
-    char === '+' ||
-    char === '*' ||
-    char === '|'
-  );
-}
