--- packages/style/src/helpers/isVariable.ts
@@ -1,3 +1,0 @@
-export default function isVariable(value: string): boolean {
-  return value.slice(0, 2) === '--';
-}
